+++ apps/backend/src/modules/contacts/contacts.module.ts
@@ -1,11 +1,13 @@
 import { Module } from '@nestjs/common';
+import { ContactImportsController } from './contact-imports.controller';
+import { ContactImportsService } from './contact-imports.service';
 import { ContactsController } from './contacts.controller';
 import { ContactsService } from './contacts.service';
 import { TagsController } from './tags.controller';
 import { TagsService } from './tags.service';
 
 @Module({
-  controllers: [ContactsController, TagsController],
-  providers: [ContactsService, TagsService],
+  controllers: [ContactImportsController, ContactsController, TagsController],
+  providers: [ContactImportsService, ContactsService, TagsService],
 })
 export class ContactsModule {}